Hi all.  This one builds and has all the data/nbin dependencies in the Models folder of the OpenNLP project.  This version doesn't yet do away with the need for separate data files and a data path.  So far now copy the files and folder structure from the Models folder and/or reference it directly in the app.config file.
My next step will be to put the data in as resources so, even though the dll will be huge, you can get a single reference to an OpenNLP instance. This can be handy when deploying web services where you don't want to or can't easily configure file paths and directories.